Refactor Ship and related classes to improve health management and streamline debug output
